refactor Feed subcomponents to separate files
