+++ apps/web/src/data/farcaster/client.ts
@@ -3,7 +3,7 @@ import { getSSLHubRpcClient } from '@farcaster/hub-nodejs'
 import { FARCASTER_HUB_RPC } from 'src/constants/farcasterHub'
 
 export const farcasterClient = () => {
-  const client = getSSLHubRpcClient(FARCASTER_HUB_RPC)
+  const client = getSSLHubRpcClient(FARCASTER_HUB_RPC as string)
 
   return client
 }
