+++ apps/web/src/modules/dao/components/Feed/CardSkeleton.tsx
@@ -1,0 +1,13 @@
+import { Box } from '@zoralabs/zord'
+
+import { cardSkeleton } from './Feed.css'
+
+export const CardSkeleton = () => (
+  <Box
+    className={cardSkeleton}
+    borderRadius="normal"
+    backgroundColor="background2"
+    style={{ height: '8rem', minHeight: '8rem' }}
+    mb="x3"
+  />
+)
